# docs/swagger.yaml
basePath: /api/v1
definitions:
  apperr.Code:
    enum:
    - core/bad_request
    - core/unauthorized
    - core/forbidden
    - core/internal_error
    type: string
    x-enum-varnames:
    - CodeBadRequest
    - CodeUnauthorized
    - CodeForbidden
    - CodeInternal
  apperr.Field:
    enum:
    - request
    type: string
    x-enum-varnames:
    - FieldRequest
  apperr.Rule:
    enum:
    - required
    - too_long
    - cycle
    - max_hierarchy_depth
    - invalid_format
    - too_short
    - duplicate
    - mismatch
    - forbidden
    - invalid_state
    - not_found
    type: string
    x-enum-varnames:
    - RuleRequired
    - RuleTooLong
    - RuleCycle
    - RuleMaxHierarchy
    - RuleInvalidFormat
    - RuleTooShort
    - RuleDuplicate
    - RuleMismatch
    - RuleForbidden
    - RuleInvalidState
    - RuleNotFound
  apperr.Violation:
    properties:
      field:
        $ref: '#/definitions/apperr.Field'
      params:
        additionalProperties: {}
        type: object
      rule:
        $ref: '#/definitions/apperr.Rule'
    type: object
  apperr.appError:
    properties:
      code:
        $ref: '#/definitions/apperr.Code'
      message:
        description: Message for user
        type: string
      violations:
        items:
          $ref: '#/definitions/apperr.Violation'
        type: array
    type: object
  auth.RefreshToken:
    properties:
      session_id:
        type: string
      token:
        type: string
    type: object
  auth.Role:
    enum:
    - admin
    - read
    - write
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleRead
    - RoleWrite
  auth.Session:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      session_version:
        type: integer
      user_id:
        type: string
    type: object
  auth.Tokens:
    properties:
      access_token:
        type: string
      refresh_token:
        $ref: '#/definitions/auth.RefreshToken'
    type: object
  auth.UserRole:
    properties:
      entity_id:
        type: string
      role:
        $ref: '#/definitions/auth.Role'
      user_id:
        type: string
    type: object
...
  http.LoginInput:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
info:
  contact: {}
  description: Demo wiki backend in Go
  title: EasyGoDocs API
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate user and get tokens
      parameters:
      - description: credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.Tokens'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperr.appError'
      summary: Login
      tags:
      - auth
  /refresh:
    post:
      consumes:
      - application/json
      description: Refreshes the access and refresh tokens using a valid refresh token
      parameters:
      - description: Refresh token payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshToken'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.Tokens'
        default:
          description: Error
          schema:
            $ref: '#/definitions/apperr.appError'
      summary: Refresh access token
      tags:
      - auth
...
  /roles:
    delete:
      consumes:
      - application/json
      description: Deletes a user role assignment for an entity. Requires admin privileges.
      parameters:
      - description: User role payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.UserRole'
      responses:
        "204":
          description: No Content
        default:
          description: Error
          schema:
            $ref: '#/definitions/apperr.appError'
      security:
      - BearerAuth: []
      summary: Remove role from user
      tags:
      - roles
    get:
      description: Returns a list of roles for the specified user ID. Requires admin
        privileges or self-access.
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/auth.UserRole'
            type: array
        default:
          description: Error
          schema:
            $ref: '#/definitions/apperr.appError'
      security:
      - BearerAuth: []
      summary: List roles assigned to a user
      tags:
      - roles
    post:
      consumes:
      - application/json
      description: Adds a role for a user in relation to an entity. Requires admin
        privileges.
      parameters:
      - description: User role payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.UserRole'
      responses:
        "204":
          description: No Content
        default:
          description: Error
          schema:
            $ref: '#/definitions/apperr.appError'
      security:
      - BearerAuth: []
      summary: Assign role to user
      tags:
      - roles
  /sessions:
    delete:
      description: Deletes all active sessions for the specified user. Requires admin
        privileges or self-access.
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        default:
          description: Error
          schema:
            $ref: '#/definitions/apperr.appError'
      security:
      - BearerAuth: []
      summary: Delete all sessions for user
      tags:
      - sessions
    get:
      description: Returns all active sessions for the specified user. Requires admin
        privileges or self-access.
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/auth.Session'
            type: array
        default:
          description: Error
          schema:
            $ref: '#/definitions/apperr.appError'
      security:
      - BearerAuth: []
      summary: List sessions by user ID
      tags:
      - sessions
  /sessions/{session_id}:
    delete:
      description: Deletes a specific session for a given user. Requires admin privileges
        or self-access.
      parameters:
      - description: Session ID
        in: path
        name: session_id
        required: true
        type: string
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        default:
          description: Error
          schema:
            $ref: '#/definitions/apperr.appError'
      security:
      - BearerAuth: []
      summary: Delete session by ID
      tags:
      - sessions
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"